WIP: Modularize properly (build errors)
