+++ ivy-core/build.gradle.kts
@@ -8,9 +8,6 @@ android {
 }
 
 dependencies {
-    implementation(projects.ivyResources) // TODO: Get rid of
-    implementation(projects.tempOldDesign) // TODO: Get rid of
-
     implementation(libs.datastore)
     implementation(libs.bundles.ktor)
 }
