+++ ivy-core/src/main/java/com/ivy/core/datamodel/Settings.kt
@@ -1,7 +1,7 @@
 package com.ivy.core.datamodel
 
 import androidx.compose.runtime.Immutable
-import com.ivy.design.l0_system.Theme
+import com.ivy.core.datamodel.legacy.Theme
 import com.ivy.core.db.entity.SettingsEntity
 import java.math.BigDecimal
 import java.util.UUID
