+++ ivy-core/src/main/java/com/ivy/core/db/entity/CategoryEntity.kt
@@ -1,12 +1,10 @@
 package com.ivy.core.db.entity
 
 import androidx.annotation.Keep
-import androidx.compose.ui.graphics.toArgb
 import androidx.room.Entity
 import androidx.room.PrimaryKey
 import com.ivy.core.datamodel.Category
 import com.ivy.core.kotlinxserilzation.KSerializerUUID
-import com.ivy.design.l0_system.Ivy
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 import java.util.*
@@ -18,7 +16,7 @@ data class CategoryEntity(
     @SerialName("name")
     val name: String,
     @SerialName("color")
-    val color: Int = Ivy.toArgb(),
+    val color: Int,
     @SerialName("icon")
     val icon: String? = null,
     @SerialName("orderNum")
