+++ ivy-navigation/build.gradle.kts
@@ -8,4 +8,5 @@ android {
 
 dependencies {
     implementation(projects.ivyCore)
+    implementation(projects.ivyDesign)
 }
