+++ ivy-navigation/src/main/java/com/ivy/navigation/IvyPreview.kt
@@ -1,8 +1,7 @@
-package com.ivy.design.system
+package com.ivy.navigation
 
 import androidx.compose.runtime.Composable
-import com.ivy.navigation.Navigation
-import com.ivy.navigation.NavigationRoot
+import com.ivy.design.system.IvyMaterial3Theme
 
 @Composable
 fun IvyPreview(
