+++ screen-home/src/main/java/com/ivy/home/HomeState.kt
@@ -2,7 +2,6 @@ package com.ivy.home
 
 import androidx.compose.runtime.Immutable
 import com.ivy.core.datamodel.TransactionHistoryItem
-import com.ivy.design.l0_system.Theme
 import com.ivy.home.customerjourney.CustomerJourneyCardModel
 import com.ivy.legacy.IvyWalletCtx
 import com.ivy.legacy.data.AppBaseData
