+++ screen-home/src/main/java/com/ivy/home/HomeViewModel.kt
@@ -3,7 +3,7 @@ package com.ivy.home
 import com.ivy.core.datamodel.Account
 import com.ivy.core.datamodel.Settings
 import com.ivy.core.datamodel.Transaction
-import com.ivy.design.l0_system.Theme
+import com.ivy.core.datamodel.legacy.Theme
 import com.ivy.frp.fixUnit
 import com.ivy.frp.then
 import com.ivy.frp.thenInvokeAfter
