+++ screen-onboarding/src/main/java/com/ivy/onboarding/viewmodel/OnboardingViewModel.kt
@@ -3,14 +3,14 @@ package com.ivy.onboarding.viewmodel
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
+import com.ivy.core.datamodel.Account
+import com.ivy.core.datamodel.Category
+import com.ivy.core.datamodel.Settings
+import com.ivy.core.datamodel.legacy.Theme
 import com.ivy.core.db.read.AccountDao
 import com.ivy.core.db.read.CategoryDao
 import com.ivy.core.db.read.SettingsDao
 import com.ivy.core.db.write.SettingsWriter
-import com.ivy.core.datamodel.Account
-import com.ivy.core.datamodel.Category
-import com.ivy.core.datamodel.Settings
-import com.ivy.design.l0_system.Theme
 import com.ivy.frp.test.TestIdlingResource
 import com.ivy.legacy.IvyWalletCtx
 import com.ivy.legacy.LogoutLogic
