+++ screen-settings/src/main/java/com/ivy/settings/SettingsViewModel.kt
@@ -5,10 +5,10 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.ivy.core.RootScreen
+import com.ivy.core.datamodel.legacy.Theme
 import com.ivy.core.db.read.SettingsDao
 import com.ivy.core.db.write.SettingsWriter
 import com.ivy.core.util.refreshWidget
-import com.ivy.design.l0_system.Theme
 import com.ivy.frp.monad.Res
 import com.ivy.frp.test.TestIdlingResource
 import com.ivy.legacy.IvyWalletCtx
