+++ screen-transaction/src/main/java/com/ivy/transaction/EditTransactionScreen.kt
@@ -22,6 +22,7 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.focus.FocusRequester
+import androidx.compose.ui.graphics.toArgb
 import androidx.compose.ui.layout.onGloballyPositioned
 import androidx.compose.ui.layout.positionInParent
 import androidx.compose.ui.res.stringResource
@@ -33,6 +34,7 @@ import androidx.lifecycle.viewmodel.compose.viewModel
 import com.ivy.core.datamodel.Account
 import com.ivy.core.datamodel.Category
 import com.ivy.core.db.entity.TransactionType
+import com.ivy.design.l0_system.Orange
 import com.ivy.design.l0_system.UI
 import com.ivy.design.l0_system.style
 import com.ivy.design.utils.hideKeyboard
@@ -606,7 +608,7 @@ private fun Preview() {
             dateTime = timeNowLocal(),
             description = null,
             category = null,
-            account = Account(name = "phyre"),
+            account = Account(name = "phyre", Orange.toArgb()),
             toAccount = null,
             amount = 0.0,
             dueDate = null,
