+++ screen-transactions/src/main/java/com/ivy/transactions/ItemStatisticScreen.kt
@@ -34,16 +34,16 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.lifecycle.viewmodel.compose.viewModel
-import com.ivy.legacy.Constants
-import com.ivy.core.db.entity.TransactionType
 import com.ivy.core.datamodel.Account
 import com.ivy.core.datamodel.Category
 import com.ivy.core.datamodel.Transaction
 import com.ivy.core.datamodel.TransactionHistoryItem
+import com.ivy.core.datamodel.legacy.Theme
+import com.ivy.core.db.entity.TransactionType
 import com.ivy.core.util.stringRes
-import com.ivy.design.l0_system.Theme
 import com.ivy.design.l0_system.UI
 import com.ivy.design.l0_system.style
+import com.ivy.legacy.Constants
 import com.ivy.legacy.IvyWalletPreview
 import com.ivy.legacy.data.AppBaseData
 import com.ivy.legacy.data.DueSection
