+++ temp-legacy-code/src/main/java/com/ivy/legacy/IvyComposeApp.kt
@@ -13,12 +13,12 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.platform.LocalView
 import com.ivy.core.RootScreen
+import com.ivy.core.datamodel.legacy.Theme
 import com.ivy.core.util.appContext
 import com.ivy.design.IvyContext
 import com.ivy.design.api.IvyDesign
 import com.ivy.design.api.ivyContext
 import com.ivy.design.api.systems.IvyWalletDesign
-import com.ivy.design.l0_system.Theme
 import com.ivy.design.l0_system.UI
 import com.ivy.design.utils.IvyPreview
 import com.ivy.navigation.Navigation
