+++ temp-legacy-code/src/main/java/com/ivy/legacy/data/model/LastNTimeRange.kt
@@ -2,6 +2,8 @@ package com.ivy.legacy.data.model
 
 import androidx.compose.runtime.Immutable
 import com.ivy.core.datamodel.IntervalType
+import com.ivy.legacy.forDisplay
+import com.ivy.legacy.incrementDate
 import com.ivy.legacy.utils.timeNowUTC
 import java.time.LocalDateTime
