+++ temp-legacy-code/src/main/java/com/ivy/legacy/legacy/ui/theme/components/IntervalPickerRow.kt
@@ -28,6 +28,7 @@ import com.ivy.design.l0_system.style
 import com.ivy.resources.R
 import com.ivy.core.datamodel.IntervalType
 import com.ivy.legacy.IvyWalletComponentPreview
+import com.ivy.legacy.forDisplay
 import com.ivy.wallet.ui.theme.Gradient
 import com.ivy.wallet.ui.theme.GradientIvy
 import com.ivy.wallet.ui.theme.White
