+++ temp-legacy-code/src/main/java/com/ivy/legacy/legacy/ui/theme/modal/LoanRecordModal.kt
@@ -29,14 +29,24 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.input.TextFieldValue
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.ivy.legacy.IvyWalletPreview
-import com.ivy.legacy.ivyWalletCtx
+import com.ivy.core.datamodel.Account
+import com.ivy.core.datamodel.LoanRecord
 import com.ivy.design.l0_system.UI
 import com.ivy.design.l0_system.style
 import com.ivy.frp.test.TestingContext
+import com.ivy.legacy.IvyWalletPreview
+import com.ivy.legacy.ivyWalletCtx
+import com.ivy.legacy.legacy.ui.theme.modal.ModalNameInput
+import com.ivy.legacy.utils.convertUTCtoLocal
+import com.ivy.legacy.utils.formatLocalTime
+import com.ivy.legacy.utils.formatNicely
+import com.ivy.legacy.utils.getDefaultFIATCurrency
+import com.ivy.legacy.utils.getTrueDate
+import com.ivy.legacy.utils.onScreenStart
+import com.ivy.legacy.utils.selectEndTextFieldValue
+import com.ivy.legacy.utils.thenIf
+import com.ivy.legacy.utils.timeNowUTC
 import com.ivy.resources.R
-import com.ivy.core.datamodel.Account
-import com.ivy.core.datamodel.LoanRecord
 import com.ivy.wallet.domain.deprecated.logic.model.CreateAccountData
 import com.ivy.wallet.domain.deprecated.logic.model.CreateLoanRecordData
 import com.ivy.wallet.domain.deprecated.logic.model.EditLoanRecordData
@@ -49,15 +59,6 @@ import com.ivy.wallet.ui.theme.modal.edit.AccountModal
 import com.ivy.wallet.ui.theme.modal.edit.AccountModalData
 import com.ivy.wallet.ui.theme.modal.edit.AmountModal
 import com.ivy.wallet.ui.theme.toComposeColor
-import com.ivy.legacy.utils.convertUTCtoLocal
-import com.ivy.legacy.utils.formatLocalTime
-import com.ivy.legacy.utils.formatNicely
-import com.ivy.legacy.utils.getDefaultFIATCurrency
-import com.ivy.legacy.utils.getTrueDate
-import com.ivy.legacy.utils.onScreenStart
-import com.ivy.legacy.utils.selectEndTextFieldValue
-import com.ivy.legacy.utils.thenIf
-import com.ivy.legacy.utils.timeNowUTC
 import kotlinx.coroutines.launch
 import java.time.LocalDateTime
 import java.util.UUID
