+++ temp-legacy-code/src/main/java/com/ivy/legacy/ui/component/transaction/TransactionCard.kt
@@ -26,10 +26,10 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.ivy.core.db.entity.TransactionType
 import com.ivy.core.datamodel.Account
 import com.ivy.core.datamodel.Category
 import com.ivy.core.datamodel.Transaction
+import com.ivy.core.db.entity.TransactionType
 import com.ivy.design.l0_system.UI
 import com.ivy.design.l0_system.style
 import com.ivy.design.l1_buildingBlocks.IvyText
@@ -44,6 +44,7 @@ import com.ivy.legacy.utils.timeNowUTC
 import com.ivy.navigation.ItemStatisticScreen
 import com.ivy.navigation.navigation
 import com.ivy.resources.R
+import com.ivy.wallet.ui.theme.Blue
 import com.ivy.wallet.ui.theme.Gradient
 import com.ivy.wallet.ui.theme.GradientGreen
 import com.ivy.wallet.ui.theme.GradientIvy
@@ -503,8 +504,8 @@ private data class AmountTypeStyle(
 private fun PreviewUpcomingExpense() {
     IvyWalletPreview {
         LazyColumn(Modifier.fillMaxSize()) {
-            val cash = Account(name = "Cash")
-            val food = Category(name = "Food")
+            val cash = Account(name = "Cash", Green.toArgb())
+            val food = Category(name = "Food", Blue.toArgb())
 
             item {
                 TransactionCard(
