+++ temp-old-design/build.gradle.kts
@@ -9,4 +9,5 @@ android {
 dependencies {
     implementation(projects.ivyResources)
     implementation(projects.ivyDesign)
+    implementation(projects.ivyCore)
 }
