+++ temp-old-design/src/main/java/com/ivy/design/utils/Preview.kt
@@ -8,11 +8,11 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import com.ivy.core.datamodel.legacy.Theme
 import com.ivy.design.IvyContext
 import com.ivy.design.api.IvyDesign
 import com.ivy.design.api.IvyUI
 import com.ivy.design.api.systems.IvyWalletDesign
-import com.ivy.design.l0_system.Theme
 import com.ivy.design.l0_system.UI
 
 @Deprecated("Old design system. Use `:ivy-design` and Material3")
